유저 이메일 대신 아이디를 식별자로 사용 로그인할 때 ClientApp.userId에 유저 아이디 저장 유저 아이디 기반으로 채팅 수정 (서버에 보내는 Message 모델 데이터베이스에 저장하는 Message 모델 수정) 채팅방에서 유저 정보들 해싱 처리

+++ app/src/main/java/com/doubleslash/playground/retrofit/dto/response/Sign_in_responseDTO.java
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName;
 public class Sign_in_responseDTO {
     @SerializedName("result")
     @Expose
-    private Integer result;
+    private Long result;
     @SerializedName("message")
     @Expose
     private String message;
@@ -14,11 +14,11 @@ public class Sign_in_responseDTO {
     @Expose
     private String token;
 
-    public Integer getResult() {
+    public Long getResult() {
         return result;
     }
 
-    public void setResult(Integer result) {
+    public void setResult(Long result) {
         this.result = result;
     }
